Revert "Merge upstream dularion branch v1.2 into branch dev1.2 (#4)"
This reverts commit d8c6cbdd4602652bd198832298fcd06a6c023d88.
