+++ build.gradle
@@ -20,17 +20,16 @@ buildscript {
 version "1.2.1_beta"
 group "streama"
 
-
 apply plugin:"eclipse"
 apply plugin:"idea"
 apply plugin:"war"
 apply plugin:"org.grails.grails-web"
 apply plugin:"org.grails.grails-gsp"
 //apply plugin:"com.craigburke.karma"
-//apply plugin:"com.craigburke.client-dependencies"
+//apply plugin:"com.craigburke.bower-installer"
 apply plugin:"asset-pipeline"
 apply plugin:"org.grails.plugins.views-json"
-apply plugin:'org.springframework.boot'
+apply plugin: 'spring-boot'
 
 
 springBoot {
@@ -55,7 +54,7 @@ dependencyManagement {
 }
 
 dependencies {
-    assets "com.craigburke.angular:angular-template-asset-pipeline:2.3.0"
+    assets "com.craigburke.angular:angular-template-asset-pipeline:2.2.6"
     assets "com.craigburke.angular:angular-annotate-asset-pipeline:2.4.0"
     assets "com.craigburke:js-closure-wrap-asset-pipeline:1.2.0"
     compile "org.springframework.boot:spring-boot-starter-logging"
@@ -75,16 +74,13 @@ dependencies {
     compile "org.grails:grails-logging"
     compile "org.grails.plugins:cache"
     compile "org.grails:grails-plugin-gsp"
-    compile "org.grails.plugins:angular-scaffolding:1.0.1"
-    compile "org.grails.plugins:hibernate5"
-    compile "org.hibernate:hibernate-core:5.1.2.Final"
-    compile "org.hibernate:hibernate-ehcache:5.1.2.Final"
-    compile "org.grails.plugins:views-json"
-    compile "org.grails.plugins:views-json-templates"
+    compile "org.grails.plugins:hibernate4"
+    compile "org.hibernate:hibernate-ehcache"
+    compile "org.grails.plugins:views-json:1.0.12"
     console "org.grails:grails-console"
-    profile "org.grails.profiles:angular"
+    profile "org.grails.profiles:angular:3.1.6"
     runtime "com.h2database:h2"
-    runtime "com.bertramlabs.plugins:asset-pipeline-grails:2.11.6"
+    runtime "com.bertramlabs.plugins:asset-pipeline-grails:2.8.2"
     testCompile "org.grails:grails-plugin-testing"
     testCompile "org.grails.plugins:geb"
     testCompile "org.grails:grails-datastore-rest-client"
