+++ gradle.properties
@@ -1,2 +1,2 @@
-grailsVersion=3.2.6
-gradleWrapperVersion=3.0
+grailsVersion=3.1.14
+gradleWrapperVersion=2.9
