--- grails-app/assets/javascripts/streama/streama-public.js
@@ -1,15 +1,0 @@
-//= wrapped
-
-//= require /systaro/core/systaro.core
-//= require /streama/streama.translations
-
-//= require_self
-
-angular.module('streamaPublic', [
-    'systaro.core',
-    'streama.translations'
-]);
-
-
-
-
