+++ grails-app/assets/stylesheets/style.css
@@ -126,6 +126,8 @@ label {
   .form-control[disabled] {
     background-color: rgba(255, 255, 255, 0.12);
     opacity: 0.5; }
+  .form-control option {
+    color: black; }
 
 [ui-sref],
 [ng-click],
@@ -566,10 +568,20 @@ h1, h2, h3, h4, h5, h6 {
   float: left;
   padding: 6px 0;
   position: relative;
-  width: 170px;
-  margin-right: 15px; }
+  margin-right: 15px;
+  display: -webkit-box;
+  display: -moz-box;
+  display: -ms-flexbox;
+  display: -webkit-flex;
+  display: flex;
+  -webkit-flex-align: flex-end;
+  -ms-flex-align: flex-end;
+  -webkit-align-items: flex-end;
+  align-items: flex-end; }
   .logo img {
-    width: 130px; }
+    max-width: 130px;
+    max-height: 31px;
+    margin-right: 10px; }
   .logo i {
     font-size: 0.85em; }
   .logo .spinner {
@@ -583,9 +595,6 @@ h1, h2, h3, h4, h5, h6 {
   display: block; }
 
 .version {
-  position: absolute;
-  top: 22px;
-  left: 137px;
   color: #B0DDEF;
   font-size: 10px; }
 
@@ -1441,12 +1450,65 @@ streama-video-player.nocursor {
   position: absolute;
   top: 50%;
   left: 50%;
-  width: auto; }
+  width: auto;
+  margin-left: -48px; }
   .player-wrapper .spinner > div {
     background: #B0DDEF;
     width: 20px;
     height: 20px;
     margin-right: 12px; }
+
+#player-menu-subtitle-selector {
+  background-color: #262626;
+  position: absolute;
+  bottom: 100%;
+  margin-bottom: 12px;
+  width: 300px;
+  right: -2.7em;
+  overflow: hidden;
+  box-shadow: 0 0 5px rgba(0, 0, 0, 0.52); }
+  #player-menu-subtitle-selector .subtitle-selector-container {
+    width: 200%; }
+    #player-menu-subtitle-selector .subtitle-selector-container h2 {
+      padding: 12px; }
+  #player-menu-subtitle-selector .subtitle-list {
+    margin: 0;
+    padding: 0;
+    border-top: 1px solid rgba(255, 255, 255, 0.09);
+    font-family: "Lato","Helvetica Neue",Helvetica,Arial,sans-serif;
+    font-size: 14px;
+    font-weight: 100;
+    box-shadow: 0px -1px 0 0 rgba(0, 0, 0, 0.46);
+    max-height: 300px;
+    overflow: auto; }
+    #player-menu-subtitle-selector .subtitle-list .flex-wrapper {
+      display: -webkit-box;
+      display: -moz-box;
+      display: -ms-flexbox;
+      display: -webkit-flex;
+      display: flex; }
+      #player-menu-subtitle-selector .subtitle-list .flex-wrapper:after {
+        content: "";
+        display: table;
+        clear: both; }
+    #player-menu-subtitle-selector .subtitle-list li {
+      color: white;
+      list-style: none;
+      padding: 0 12px;
+      border-top: 1px solid rgba(0, 0, 0, 0.27);
+      border-bottom: 1px solid rgba(255, 255, 255, 0.08); }
+      #player-menu-subtitle-selector .subtitle-list li:hover {
+        background: rgba(0, 0, 0, 0.2); }
+      #player-menu-subtitle-selector .subtitle-list li.active {
+        background: rgba(0, 0, 0, 0.2); }
+      #player-menu-subtitle-selector .subtitle-list li.current {
+        background: #1f779a; }
+        #player-menu-subtitle-selector .subtitle-list li.current .episode-play {
+          display: none; }
+      #player-menu-subtitle-selector .subtitle-list li.no-files {
+        opacity: 0.7; }
+  #player-menu-subtitle-selector .subtitle-list li {
+    padding: 12px; }
 
 .player-wrapper button.button {
   padding: .5em 2em;
@@ -1590,7 +1652,7 @@ streama-video-player.nocursor {
       color: #cacaca; }
       .player-wrapper .player-control-bar .player-control-button:before:hover {
         color: #fff; }
-    .player-wrapper .player-control-bar .player-control-button.inactive {
+    .player-wrapper .player-control-bar .player-control-button.inactive:before {
       opacity: 0.3; }
   .player-wrapper .player-control-bar.player-control-divider {
     -webkit-order: 5;
