+++ grails-app/conf/application.yml
@@ -58,8 +58,6 @@ grails:
     spring:
         transactionManagement:
             proxies: false
-
-
 info:
     app:
         name: '@info.app.name@'
@@ -111,13 +109,8 @@ grails:
         defaultScope: singleton
     converters:
         encoding: UTF-8
-    cors:
-        enabled: true
-        mappings:
-            /api/**: {}
-            /login/**: {}
-            /dash/**: {}
-            /player/**: {}
+
+
 
 grails.controllers.upload.maxFileSize: 10000000000000 # 10TB
 grails.controllers.upload.maxRequestSize: 10000000000000 # 10TB
