+++ grails-app/domain/streama/Episode.groovy
@@ -27,6 +27,7 @@ class Episode extends Video{
 
   }
 
+
   def getMovieDbMeta(){
     theMovieDbService.getEpisodeMeta(this.show.apiId, this.season_number, this.episode_number)
   }
