+++ grails-app/domain/streama/TvShow.groovy
@@ -56,13 +56,7 @@ class TvShow {
   }
 
   def getFullTvShowMeta(){
-    try{
-      return theMovieDbService.getFullTvShowMeta(this.apiId)
-    }catch (e){
-      log.warn("couldnt get FullTvShowMeta for ${this.apiId}")
-      log.warn(e.message)
-      return null
-    }
+    return theMovieDbService.getFullTvShowMeta(this.apiId)
   }
