+++ grails-app/services/streama/VideoService.groovy
@@ -13,19 +13,4 @@ class VideoService {
     video.files?.each{ fileService.fullyRemoveFile(it) }
     video.save failOnError: true, flush: true
   }
-
-
-  public static List<ViewingStatus> listContinueWatching(User currentUser) {
-    List<ViewingStatus> continueWatching = ViewingStatus.withCriteria {
-      eq("user", currentUser)
-      video {
-        isNotEmpty("files")
-        ne("deleted", true)
-      }
-//      eq("completed", false)
-      order("lastUpdated", "desc")
-    }
-    return continueWatching
-  }
-
 }
