--- grails-app/views/dash/listContinueWatching.gson
@@ -1,8 +1,0 @@
-import streama.ViewingStatus
-
-model {
-  Iterable<ViewingStatus> viewingStatusList
-}
-
-
-json tmpl.viewingStatus(viewingStatusList ?: [])
