+++ grails-app/views/invite/index.gsp
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en" class="no-js" ng-app="streamaPublic">
+<html lang="en" class="no-js" ng-app="streama">
   <head>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
     <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
@@ -66,7 +66,7 @@
 
 
   <asset:javascript src="vendor.js" />
-  <asset:javascript src="streama/streama-public.js" />
+  <asset:javascript src="streama/streama.js" />
 
   </body>
 </html>
